mod: navigate 정리
